Include simple worker example
This worker uses some other packages such as split, safe, and text;
these are included in the cabal file. This worker reads in a csv file of
employees and builds an Employee struct out of each line in the file.
The idea is to demonstrate that each employee could then be filtered for
all employees above a given age or who are software developers, etc.